+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project follows Semantic Versioning.
 
+## [Unreleased]
+- Added multiple selection support for the AppShellSelectComponent and AppShellFiltersComponent.
+
 ## [19.0.15]
 
 Library version: `@opendevstack/ngx-appshell@19.0.15`